requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Data Science & ML
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
joblib==1.3.2

# Deep Learning
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1

# Explainability
shap==0.43.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Data Validation
great-expectations==0.18.4

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
